backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
crewai==0.35.8
crewai-tools==0.2.6
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai
pydantic<2.8
python-docx
pypdf2
numexpr
faiss-cpu
motor[srv]
python-dotenv